cleaned up index.js
